refactor(front): rename folders
