+++ front/src/components/atoms/AnimatedPulseAtom.tsx
