+++ front/src/components/atoms/OmnilogLogo.tsx
