+++ front/src/components/atoms/TheodoLogo.tsx
