+++ front/src/features/LogFilter/FilterHeader.tsx
@@ -1,8 +1,8 @@
 "use client";
 import { faMagnifyingGlass } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { CardAtom } from "../../atomic/atoms/CardAtom";
-import { TextInput } from "../../atomic/atoms/TextInput";
+import { CardAtom } from "../../components/atoms/CardAtom";
+import { TextInput } from "../../components/atoms/TextInput";
 import { TimeDropdown } from "./TimeDropdown";
 
 export const FilterHeader = () => {
