+++ front/src/features/LogFilter/TimeDropdown.tsx
@@ -1,5 +1,5 @@
 import { useNavigation } from "@/hooks/useNavigation";
-import { Dropdown } from "../../atomic/molecules/Dropdown";
+import { Dropdown } from "../../components/molecules/Dropdown";
 
 export const TimeDropdown = () => {
     const { searchParams, updateSearchParam } = useNavigation();
