+++ front/src/features/LogTables/EmptyLogs.tsx
@@ -1,4 +1,4 @@
-import { CardAtom } from "@/atomic/atoms/CardAtom";
+import { CardAtom } from "@/components/atoms/CardAtom";
 import Link from "next/link";
 
 export default function EmptyLogs() {
