+++ front/src/services/__tests__/extractDataFromSearchParam.tsx
